renamed the files for the convention and updated reference
